Feature/MGMIDN-1038 sp 2 edit matches be
Merge in DA-NANG-INTERNSHIP/easy-tourney-be from feature/MGMIDN-1038-sp-2-edit-matches-be to master

Squashed commit of the following:

commit 6097ebf5d98781eefa0190fd22ae83f5d5f4b8d1
Merge: 55a65ec 0cb911d
Author: duypro432003 <vanduyluu432003@gmail.com>
Date:   Fri Dec 29 18:22:02 2023 +0700

    Merge branch 'master' into feature/MGMIDN-1038-sp-2-edit-matches-be

commit 55a65ec153c25d0cda1c8dee1f5cffc76dd17a3f
Author: duypro432003 <vanduyluu432003@gmail.com>
Date:   Fri Dec 29 18:21:01 2023 +0700

    fix

commit 9f14765534ae7227952fbb1c5f345a7895872c55
Author: duypro432003 <vanduyluu432003@gmail.com>
Date:   Fri Dec 29 17:15:23 2023 +0700

    fix bug

commit 7b3ca66e50b077fabe8c50c56e8211c88d469e25
Merge: 27eceff a0c40f5
Author: duypro432003 <vanduyluu432003@gmail.com>
Date:   Fri Dec 29 17:14:33 2023 +0700

    Merge branch 'master' into feature/MGMIDN-1038-sp-2-edit-matches-be

commit 27eceff4e56021633fce7ea5e023926ef0a052af
Author: duypro432003 <vanduyluu432003@gmail.com>
Date:   Fri Dec 29 17:09:24 2023 +0700

    fix edit match

commit fdba2b8c0db97140cf86ff3b645a2c239c9e455f
Author: duypro432003 <vanduyluu432003@gmail.com>
Date:   Wed Dec 27 16:27:17 2023 +0700

    edit match

commit 01181b14aff1d632813794af09b8766e8256edca
Author: Biên Long Nguyễn <bien.long.nguyen@mgm-tp.com>
Date:   Wed Dec 27 11:09:52 2023 +0000

    MGMIDN-1046: Show leaderboard BE

    Merge in DA-NANG-INTERNSHIP/easy-tourney-be from feature/MGMIDN-1046-sp-2-show-leaderboard-be to master

    Squashed commit of the following:

    commit 836493ccd0d87aea26ba5fa3cc882c4505412056
    Author: Long Bien <lbit@pop-os.localdomain>
    Date:   Wed Dec 27 14:01:20 2023 +0700

        MGMIDN-1046: Show leaderboard BE

commit bb88549ef5a751d7dac2cb9daae47ca17cdbde98
Author: duypro432003 <vanduyluu432003@gmail.com>
Date:   Thu Dec 28 17:27:33 2023 +0700

    edit match

commit 5af2d38b7b439421b1c3ce9b497c10b8e94ef3f9
Author: duypro432003 <vanduyluu432003@gmail.com>
Date:   Wed Dec 27 16:27:17 2023 +0700

    edit match

+++ src/main/java/com/example/easytourneybe/team/TeamController.java
@@ -46,21 +46,21 @@ public class TeamController {
         );
     }
     @PutMapping("/team/{id}")
-    public ResponseEntity<ResponseObject> updateTeam(@PathVariable Integer tournament_id, @PathVariable Integer id, @Valid @RequestBody Team team) {
+    public ResponseEntity<ResponseObject> updateTeam(@PathVariable Integer tournament_id, @PathVariable Long id, @Valid @RequestBody Team team) {
         Optional<Team> updateTeam = teamService.updateTeam(tournament_id, id, team.getTeamName().trim());
         return ResponseEntity.status(HttpStatus.OK).body(
                 new ResponseObject(true, 1, updateTeam)
         );
     }
     @DeleteMapping("/team/{id}")
-    public ResponseEntity<ResponseObject> deleteTeam(@PathVariable Integer tournament_id, @PathVariable Integer id) {
+    public ResponseEntity<ResponseObject> deleteTeam(@PathVariable Integer tournament_id, @PathVariable Long id) {
         Optional<Team> deleteTeam = teamService.deleteTeam(tournament_id, id);
         return ResponseEntity.status(HttpStatus.OK).body(
                 new ResponseObject(true, 1, deleteTeam));
     }
 
     @GetMapping("/team/{id}")
-    public ResponseEntity<ResponseObject> findTeamById(@PathVariable Integer tournament_id, @PathVariable Integer id) {
+    public ResponseEntity<ResponseObject> findTeamById(@PathVariable Integer tournament_id, @PathVariable Long id) {
         Team team = teamService.findTeamById(tournament_id, id);
         return ResponseEntity.status(HttpStatus.OK).body(
                 new ResponseObject(true, 1, team)
